dv_repair: Add full support for all recording timestamp packs
We were missing all audio auxiliary packs, as well as the binary group
packs.

However, these aren't yet known to actually show up in the subcode
DIF block, so don't actually add support for them to that DIF block
unless we actually need to later on: it would clutter the CSV file
otherwise.